debugging...

+++ src/kernels/data_backup.h
@@ -17,11 +17,11 @@ namespace starflow {
 			raft::kstatus run() override;
 		private:
 			/* Feature Extraction */
-			float flow_duration(const std::list<types::Packet>& packets) const;				//Done
-			float min_interarrival_time(const std::list<types::Packet>& packets) const;		//Done
-			float mean_interarrival_time(const std::list<types::Packet>& packets) const;	//Done
-			float max_interarrival_time(const std::list<types::Packet>& packets) const;		//Done
-			float stddev_interarrival_time(const std::list<types::Packet>& packets) const;	//Done
+			unsigned long flow_duration(const std::list<types::Packet>& packets) const;				//Done
+			unsigned long min_interarrival_time(const std::list<types::Packet>& packets) const;		//Done
+			unsigned long mean_interarrival_time(const std::list<types::Packet>& packets) const;	//Done
+			unsigned long max_interarrival_time(const std::list<types::Packet>& packets) const;		//Done
+			unsigned long stddev_interarrival_time(const std::list<types::Packet>& packets) const;	//Done
 			unsigned long min_packet_length(const std::list<types::Packet>& packets) const;			//Done
 			unsigned long mean_packet_length(const std::list<types::Packet>& packets) const;		//Done
 			unsigned long max_packet_length(const std::list<types::Packet>& packets) const;			//Done
